+++ src/migrations/versions/default/__init__.py
@@ -1,0 +1,1 @@
+"""Default-namespace Alembic revisions for the ms-markets provider."""